refactor code to prepare it for unit testing
